requirements.txt
alembic==1.10.2
altair==4.2.2
async-generator==1.10
attrs==22.2.0
blinker==1.5
branca==0.6.0
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
contourpy==1.0.7
cycler==0.11.0
decorator==5.1.1
entrypoints==0.4
et-xmlfile==1.1.0
Flask==2.2.3
Flask-Login==0.6.2
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.3
folium==0.14.0
fonttools==4.38.0
geographiclib==2.0
geopy==2.3.0
gitdb==4.0.10
GitPython==3.1.31
greenlet==2.0.2
h11==0.14.0
idna==3.4
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
kiwisolver==1.4.4
Mako==1.2.4
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.7.0
mdurl==0.1.2
numpy==1.24.3
openpyxl==3.1.1
outcome==1.2.0
packaging==23.0
pandas==1.5.3
patsy==0.5.3
Pillow==9.4.0
plotly==5.13.1
plotly-express==0.4.1
protobuf==3.20.3
pyarrow==11.0.0
pycparser==2.21
pydeck==0.8.0
Pygments==2.14.0
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.19.3
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
requests==2.28.2
rich==13.3.1
scikit-learn==1.2.2
scipy==1.10.1
selenium==4.8.2
semver==2.13.0
six==1.16.0
sklearn==0.0.post1
smmap==5.0.0
sniffio==1.3.0
sortedcontainers==2.4.0
SQLAlchemy==2.0.7
statsmodels==0.13.5
streamlit==1.12.0
tenacity==8.2.2
threadpoolctl==3.1.0
toml==0.10.2
toolz==0.12.0
tornado==6.2
trio==0.22.0
trio-websocket==0.9.2
typing_extensions==4.5.0
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.15
validators==0.20.0
watchdog==2.3.0
Werkzeug==2.2.3
wsproto==1.2.0
zipp==3.15.0